Fix various things and refactoring
