+++ src/openelm/benchmarks_diff.py
@@ -1,10 +1,8 @@
 """This is the benchmark test of diff models."""
 import functools
-import itertools
 import json
 import multiprocessing as mp
 import re
-from typing import Union
 
 import hydra
 import numpy as np
@@ -13,81 +11,14 @@ from omegaconf import OmegaConf
 from tqdm import tqdm
 from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer
 
-from openelm.benchmarks import eval_code_string, parity_reference
 from openelm.constants import SRC_PATH
 from openelm.utils.diff_eval import apply_diff, split_diff
-
-parity_test_data = {
-    i: parity_reference(*i) for i in itertools.product(range(2), repeat=4)
-}
-
-
-def eval_completions(
-    eval_results: Union[str, list[str]], task: str = "parity", timeout: int = 5
-) -> Union[int, list[int]]:
-    """
-    Evaluate (a batch of) the modified eval_results on a task.
-
-    Args:
-        eval_results: either a string or a batch of strings. The code(s) to be evaluated.
-        task: (Optional) the task to be performed.
-        timeout: (Optional) the timeout (in seconds).
-
-    Returns:
-        either the status code of the string or a list of status eval_results of the batch
-        of strings (depending on whether `eval_results` is batched).
-    """
-    if task == "parity":
-        _eval_results = (
-            eval_results if isinstance(eval_results, list) else [eval_results]
-        )
-        results = []
-        for code in _eval_results:
-            results.append(eval_code_string(code, parity_test_data, timeout))
-        if isinstance(eval_results, list):
-            # Batch evaluation returns the batch
-            return results
-        else:
-            # Single evaluation returns a single result
-            return results[0]
-    else:
-        raise ValueError(f"Unknown task: {task}")
-
-
-def mutate_code(n_bugs: int = 5, task: str = "parity") -> tuple:
-    """
-    Mutate code to create n bugs. Output the prompt in diff format.
-
-    Args:
-        n_bugs: number of bugs to introduce (from 1 to 5).
-        task: (Optional) the task to be performed.
-
-    Returns:
-        mutated_code, function_string
-    """
-    mutation_template = [
-        f"<NME> {task}.py\n<BEF> ",
-        "",  # placeholder for the context, e.g., the buggy code
-        "\n<MSG> Fixed bugs",
-    ]
-    if task == "parity":
-        variables = ["b", "b", "b", "b", 2]
-        for i in range(n_bugs):
-            variables[i] = "c" if i < 4 else 3
-        func_str = (
-            'def parity(b1,b2,b3,b4):\n    """Return binary parity of a sequence of input bits.'
-            ' Return 0 for even parity, 1 for odd parity."""\n    bit_sum = sum(['
-            "{}1,{}2,{}3,{}4])\n    return bit_sum % {}".format(*variables)
-        )
-        mutation_template[1] = func_str
-        return "".join(mutation_template), func_str
-    else:
-        raise ValueError(f"Unknown task: {task}")
+from openelm.utils.code_eval import eval_completions, mutate_code
 
 
 def benchmark_parity(cfg, model, tokenizer, device, n_bugs, temperature):
     # Prepare the mutated eval_results with cfg.n_bugs bugs.
-    mutated_str, function_str = mutate_code(n_bugs=n_bugs, task="parity")
+    mutated_str, function_str = mutate_code(n_bugs=n_bugs, task="parity", mutate_method="diff")
     mutated_encoding = tokenizer(
         [mutated_str],
         truncation=True,
@@ -180,7 +111,7 @@ def benchmark_bugs(cfg, model, tokenizer, device, n_trials_per_task):
             texts = tokenizer.batch_decode(tokens)
 
         end_of_diff = re.compile("\n[^ +-@]+")
-        section_names = set(["name", "file", "message", "diff"])
+        section_names = {"name", "file", "message", "diff"}
         for j in range(len(texts)):
             # split the diff text according to <NME>, <BEF>, <MSG>, <DFF>.
             parsed = split_diff(texts[j])
@@ -206,6 +137,7 @@ def benchmark_bugs(cfg, model, tokenizer, device, n_trials_per_task):
         f"{(corr_cnt / (len(eval_results))) * 100}%",
     )
     return (corr_cnt / (len(eval_results))) * 100
+
 
 @hydra.main(
     config_path=str(SRC_PATH / "config"),
